Added hw2 part 2 repo

+++ hw2p2/MovingThread.cpp
@@ -1,5 +1,4 @@
 #include "MovingThread.h"
-//Correct version
 
 bool busy;
 int i;
@@ -38,7 +37,7 @@ bool MovingThread::isBusy() {
 void MovingThread::run() {
     int64_t tic = 0;
     int64_t currentTic;
-    float ticLength;
+    double ticLength;
     while (!(*stopped)) {
         ticLength = timeline->getRealTicLength();
         currentTic = timeline->getTime();
@@ -58,12 +57,12 @@ void MovingThread::run() {
                 if (!(i->getType())) {
                     //Check the bounds. Round to the nearest 2 decimal places to avoid floating point errors.
                     if ((int)i->getGlobalBounds().top < bounds.y || (int)i->getGlobalBounds().top > bounds.x ) {
-                        i->setVSpeed(sf::Vector2f(0, i->getSpeed().y * -1.f));
+                        i->setVSpeed(sf::Vector2f(0, i->getSpeed().y * -1.0));
                     }
                 }
                 else {
                     if (i->getGlobalBounds().left > bounds.y || i->getGlobalBounds().left < bounds.x ) {
-                        i->setHSpeed(sf::Vector2f(i->getSpeed().x * -1.f, 0));
+                        i->setHSpeed(sf::Vector2f(i->getSpeed().x * -1.0, 0));
                     }
                 }
                 _condition_variable->notify_all();
